requirements.txt
Pillow
numpy
matplotlib
requests
